Move tests for conversion into numpy into a separate file
