+++ docs/api/frame/to_numpy.rst
@@ -2,3 +2,4 @@
 .. xmethod:: datatable.Frame.to_numpy
     :src: src/core/frame/to_numpy.cc Frame::to_numpy
     :doc: src/core/frame/to_numpy.cc doc_to_numpy
+    :tests: tests/frame/test-to-numpy.py
