+++ tests/test-dt.py
@@ -744,189 +744,6 @@ def test_to_dict():
 
 
 
-#-------------------------------------------------------------------------------
-# Test conversions into Numpy
-#-------------------------------------------------------------------------------
-
-def test_tonumpy0(numpy):
-    d0 = dt.Frame([1, 3, 5, 7, 9])
-    assert d0.stype == dt.int32
-    a0 = d0.to_numpy()
-    assert a0.shape == d0.shape
-    assert a0.dtype == numpy.dtype("int32")
-    assert a0.tolist() == [[1], [3], [5], [7], [9]]
-    a1 = numpy.array(d0)
-    assert (a0 == a1).all()
-
-
-def test_tonumpy1(numpy):
-    d0 = dt.Frame({"A": [1, 5], "B": ["helo", "you"],
-                   "C": [True, False], "D": [3.4, None]})
-    a0 = d0.to_numpy()
-    assert a0.shape == d0.shape
-    assert a0.dtype == numpy.dtype("object")
-    assert list_equals(a0.T.tolist(), d0.to_list())
-    a1 = numpy.array(d0)
-    assert (a0 == a1).all()
-
-
-def test_numpy_constructor_simple(numpy):
-    tbl = [[1, 4, 27, 9, 22], [-35, 5, 11, 2, 13], [0, -1, 6, 100, 20]]
-    d0 = dt.Frame(tbl)
-    assert d0.shape == (5, 3)
-    assert d0.stypes == (stype.int32, stype.int32, stype.int32)
-    assert d0.to_list() == tbl
-    n0 = numpy.array(d0)
-    assert n0.shape == d0.shape
-    assert n0.dtype == numpy.dtype("int32")
-    assert n0.T.tolist() == tbl
-
-
-def test_numpy_constructor_empty(numpy):
-    d0 = dt.Frame()
-    assert d0.shape == (0, 0)
-    n0 = numpy.array(d0)
-    assert n0.shape == (0, 0)
-    assert n0.tolist() == []
-
-
-def test_numpy_constructor_multi_types(numpy):
-    # Test that multi-types datatable will be promoted into a common type
-    tbl = [[1, 5, 10],
-           [True, False, False],
-           [30498, 1349810, -134308],
-           [1.454, 4.9e-23, 10000000]]
-    d0 = dt.Frame(tbl)
-    assert d0.stypes == (dt.int32, dt.bool8, dt.int32, dt.float64)
-    n0 = numpy.array(d0)
-    assert n0.dtype == numpy.dtype("float64")
-    assert n0.T.tolist() == [[1.0, 5.0, 10.0],
-                             [1.0, 0, 0],
-                             [30498, 1349810, -134308],
-                             [1.454, 4.9e-23, 1e7]]
-    assert (d0.to_numpy() == n0).all()
-
-
-def test_numpy_constructor_view(numpy):
-    d0 = dt.Frame([range(100), range(0, 1000000, 10000)])
-    d1 = d0[::-2, :]
-    assert isview(d1)
-    n1 = numpy.array(d1)
-    assert n1.dtype == numpy.dtype("int32")
-    assert n1.T.tolist() == [list(range(99, 0, -2)),
-                             list(range(990000, 0, -20000))]
-    assert (d1.to_numpy() == n1).all()
-
-
-def test_numpy_constructor_single_col(numpy):
-    d0 = dt.Frame([1, 1, 3, 5, 8, 13, 21, 34, 55])
-    assert d0.stype == dt.int32
-    n0 = numpy.array(d0)
-    assert n0.shape == d0.shape
-    assert n0.dtype == numpy.dtype("int32")
-    assert (n0 == d0.to_numpy()).all()
-
-
-def test_numpy_constructor_single_string_col(numpy):
-    d = dt.Frame(["adf", "dfkn", "qoperhi"])
-    assert d.shape == (3, 1)
-    assert d.stypes == (stype.str32, )
-    a = numpy.array(d)
-    assert a.shape == d.shape
-    assert a.dtype == numpy.dtype("object")
-    assert a.T.tolist() == d.to_list()
-    assert (a == d.to_numpy()).all()
-
-
-def test_numpy_constructor_view_1col(numpy):
-    d0 = dt.Frame({"A": [1, 2, 3, 4], "B": [True, False, True, False]})
-    d2 = d0[::2, "B"]
-    a = d2.to_numpy()
-    assert a.T.tolist() == [[True, True]]
-    assert (a == numpy.array(d2)).all()
-
-
-def test_tonumpy_with_stype(numpy):
-    """
-    Test using dt.to_numpy() with explicit `stype` argument.
-    """
-    src = [[1.5, 2.6, 5.4], [3, 7, 10]]
-    d0 = dt.Frame(src)
-    assert d0.stypes == (stype.float64, stype.int32)
-    a1 = d0.to_numpy("float32")
-    a2 = d0.to_numpy()
-    del d0
-    # Check using list_equals(), which allows a tolerance of 1e-6, because
-    # conversion to float32 resulted in loss of precision
-    assert list_equals(a1.T.tolist(), src)
-    assert a2.T.tolist() == src
-    assert a1.dtype == numpy.dtype("float32")
-    assert a2.dtype == numpy.dtype("float64")
-
-
-def test_tonumpy_with_NAs1(numpy):
-    src = [1, 5, None, 187, None, 103948]
-    d0 = dt.Frame(src)
-    a0 = d0.to_numpy()
-    assert a0.T.tolist() == [src]
-
-
-def test_tonumpy_with_NAs2(numpy):
-    src = [[2.3, 11.89, None, math.inf], [4, None, None, -12]]
-    d0 = dt.Frame(src)
-    a0 = d0.to_numpy()
-    assert a0.T.tolist() == src
-
-
-def test_tonumpy_with_NAs3(numpy):
-    src = ["faa", None, "", "hooray", None]
-    d0 = dt.Frame(src)
-    a0 = d0.to_numpy()
-    assert a0.T.tolist() == [src]
-
-
-def test_tonumpy_with_NAs4(numpy):
-    src = [True, False, None]
-    d0 = dt.Frame(src)
-    a0 = d0.to_numpy()
-    assert a0.dtype == numpy.dtype("bool")
-    assert a0.T.tolist() == [src]
-
-
-@pytest.mark.parametrize("seed", [random.getrandbits(32)])
-def test_tonumpy_with_NAs_random(seed, numpy):
-    random.seed(seed)
-    n = int(random.expovariate(0.001) + 1)
-    m = int(random.expovariate(0.2) + 1)
-    data = [None] * m
-    for j in range(m):
-        threshold = 0.1 * (m + 1);
-        vec = [random.random() for i in range(n)]
-        for i, x in enumerate(vec):
-            if x < threshold:
-                vec[i] = None
-        data[j] = vec
-    DT = dt.Frame(data)
-    ar = DT.to_numpy()
-    assert ar.T.tolist() == data
-
-
-@pytest.mark.usefixtures("numpy")
-def test_tonumpy_with_NAs_view():
-    # See issue #1738
-    X = dt.Frame(A=[5.7, 2.3, None, 4.4, 9.8, None])[1:, :]
-    a = X.to_numpy()
-    assert a.tolist() == [[2.3], [None], [4.4], [9.8], [None]]
-
-
-@pytest.mark.usefixtures("numpy")
-def test_tonumpy_issue2050():
-    n = 1234
-    DT = dt.Frame(A=[1,2,None,4,5], B=range(5), C=[4, None, None, None, 4], stype=int)
-    DT = dt.repeat(DT[:, ["A", "B", "C"]], n)
-    assert DT.sum().to_list() == [[12*n], [10*n], [8*n]]
-    assert DT.to_numpy().sum() == 30*n
-
 
 
 #-------------------------------------------------------------------------------
